correctifs de page d'accueil

+++ pockedex/src/composants/MonPokedex.js
@@ -1,41 +1,71 @@
 import React, { useState, useEffect } from 'react';
-import { Link } from 'react-router-dom'; // Importer le composant de lien depuis React Router
-import Header from './Header'; // Importer le composant Header
+import Header from './Header';
+import './css/MonPokedex.css';
 
 const MonPokedex = () => {
   const [pokemons, setPokemons] = useState([]);
+  const [searchTerm, setSearchTerm] = useState('');
 
   useEffect(() => {
-    // Charger les données depuis le localStorage
     const savedPokemons = JSON.parse(localStorage.getItem('pokemons')) || [];
     setPokemons(savedPokemons);
   }, []);
 
   const handleRemovePokemon = (pokemonName) => {
-    // Supprimer le Pokémon du localStorage
-    const updatedPokemons = pokemons.filter(pokemon => pokemon.name !== pokemonName);
-    setPokemons(updatedPokemons);
-    localStorage.setItem('pokemons', JSON.stringify(updatedPokemons));
+    const confirmRemove = window.confirm(`Êtes-vous sûr de supprimer ${pokemonName} de votre Pokédex ?`);
+    if (confirmRemove) {
+      const updatedPokemons = pokemons.filter(pokemon => pokemon.name !== pokemonName);
+      setPokemons(updatedPokemons);
+      localStorage.setItem('pokemons', JSON.stringify(updatedPokemons));
+    }
   };
 
+  const handleClearPokedex = () => {
+    const confirmClear = window.confirm("Êtes-vous sûr de vider votre Pokédex ?");
+    if (confirmClear) {
+      localStorage.removeItem('pokemons');
+      setPokemons([]);
+    }
+  };
+
+  const handleSearch = (event) => {
+    setSearchTerm(event.target.value);
+  };
+
+  const filteredPokemons = pokemons.filter((pokemon) =>
+    pokemon.name.toLowerCase().includes(searchTerm.toLowerCase())
+  );
+
   return (
-    <div>
-      <Header /> {/* Utilisation du composant Header */}
+    <div className="pokedex-container">
+      <Header />
       <h1>Mon Pokédex</h1>
-      <div className="pokemon-list-container"> {/* Utilisation d'une classe CSS pour le conteneur de liste */}
-        {pokemons.map((pokemon, index) => (
-          <div key={index} className="pokemon-card"> {/* Utilisation d'une classe CSS pour chaque carte de Pokémon */}
-            <img src={pokemon.sprites.front_default} alt={pokemon.name} className="pokemon-image" /> {/* Utilisation d'une classe CSS pour l'image */}
-            <div className="pokemon-details"> {/* Utilisation d'une classe CSS pour les détails du Pokémon */}
+      <div className="search-container">
+        <input
+          type="text"
+          placeholder="Rechercher un Pokémon"
+          value={searchTerm}
+          onChange={handleSearch}
+          className="search-input"
+        />
+      </div>
+      <div className="pokemon-list-container">
+        {filteredPokemons.map((pokemon, index) => (
+          <div key={index} className="pokemon-card">
+            <img src={pokemon.sprites.front_default} alt={pokemon.name} className="pokemon-image" />
+            <div className="pokemon-details">
               <p><strong>Nom:</strong> {pokemon.name}</p>
               <p><strong>Numéro:</strong> {pokemon.id}</p>
               <p><strong>Types:</strong> {pokemon.types.map(type => type.type.name).join(', ')}</p>
-              <button onClick={() => handleRemovePokemon(pokemon.name)}>Supprimer</button>
+              <a href={`/pokedex/${pokemon.name}`}><button className="detail-button">Voir le détail</button></a>
+              <button className="remove-button" onClick={() => handleRemovePokemon(pokemon.name)}>Supprimer</button>
             </div>
           </div>
         ))}
       </div>
-      <button onClick={() => localStorage.removeItem('pokemons')}>Vider le Pokédex</button>
+      <div className="center">
+        <button className="clear-button" onClick={handleClearPokedex}>Vider le Pokédex</button>
+      </div>
     </div>
   );
 };
